+++ .gitignore
@@ -4,6 +4,8 @@ build/
 !gradle/wrapper/gradle-wrapper.jar
 !**/src/main/**/build/
 !**/src/test/**/build/
+!**/src/main/resources/application-prod.yml
+
 
 ### STS ###
 .apt_generated
@@ -35,3 +37,4 @@ out/
 
 ### VS Code ###
 .vscode/
+/src/main/resources/application-prod.yml
