+++ src/main/java/com/example/jubging/DTO/LoginDTO.java
@@ -1,0 +1,24 @@
+package com.example.jubging.DTO;
+
+import com.example.jubging.Model.User;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.springframework.security.crypto.password.PasswordEncoder;
+
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class LoginDTO {
+    private String userId;
+    private String password;
+
+    public User toUser(PasswordEncoder passwordEncoder) {
+        return User.builder()
+                .userId(userId)
+                .password(passwordEncoder.encode(password))
+                .build();
+    }
+}
