+++ src/main/java/com/example/jubging/DTO/TokenDTO.java
@@ -1,0 +1,15 @@
+package com.example.jubging.DTO;
+
+import lombok.*;
+
+@Getter
+@Setter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class TokenDTO {
+    private String grantType;
+    private String accessToken;
+    private String refreshToken;
+    private Long accessTokenExpireDate;
+}
