+++ src/main/java/com/example/jubging/Exception/CEmailLoginFailedException.java
@@ -1,0 +1,16 @@
+package com.example.jubging.Exception;
+
+// 로그인 예외처리
+public class CEmailLoginFailedException extends RuntimeException {
+    public CEmailLoginFailedException() {
+        super();
+    }
+
+    public CEmailLoginFailedException(String message) {
+        super(message);
+    }
+
+    public CEmailLoginFailedException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
