+++ src/main/java/com/example/jubging/Exception/CUserNotFoundException.java
@@ -1,0 +1,16 @@
+package com.example.jubging.Exception;
+
+public class CUserNotFoundException extends RuntimeException{
+
+    public CUserNotFoundException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public CUserNotFoundException(String message) {
+        super(message);
+    }
+
+    public CUserNotFoundException() {
+        super();
+    }
+}
