+++ src/main/java/com/example/jubging/Model/Gender.java
@@ -1,0 +1,8 @@
+package com.example.jubging.Model;
+
+public enum Gender {
+// Male - 0
+// Female - 1
+// None - 2
+    MALE, FEMALE, NONE
+}
