+++ src/main/java/com/example/jubging/Model/PloggingRecords.java
@@ -1,0 +1,34 @@
+package com.example.jubging.Model;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+import java.time.LocalDateTime;
+
+@Builder
+@Entity
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Table(name = "plogging_records")
+public class PloggingRecords {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long number;
+
+    @Column(name = "user_id", nullable = false)
+    private String userId;
+
+    @Column(name = "date", nullable = false)
+    private LocalDateTime date;
+
+    @Column(name = "distance", nullable = false)
+    private Double distance;
+
+    @Column(name = "activity_time", nullable = false)
+    private String activityTime;
+
+}
