+++ src/main/java/com/example/jubging/Repository/TokenRepository.java
@@ -1,0 +1,10 @@
+package com.example.jubging.Repository;
+
+import com.example.jubging.Model.RefreshToken;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface TokenRepository extends JpaRepository<RefreshToken, Long> {
+    Optional<RefreshToken> findByTokenKey(Long tokenKey);
+}
