+++ src/main/java/com/example/jubging/Response/SingleResult.java
@@ -1,0 +1,10 @@
+package com.example.jubging.Response;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class SingleResult<T> extends CommonResult{
+    private T data;
+}
