+++ src/main/java/com/example/jubging/Service/CustomUserDetailService.java
@@ -1,0 +1,26 @@
+package com.example.jubging.Service;
+
+import com.example.jubging.Exception.CUserNotFoundException;
+import com.example.jubging.Repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+
+@RequiredArgsConstructor
+@Service
+public class CustomUserDetailService implements UserDetailsService {
+
+    private final UserRepository userRepository;
+
+    @Override
+    @Transactional(readOnly = true)
+    public UserDetails loadUserByUsername(String userPk) throws UsernameNotFoundException {
+        return userRepository.findById(Long.parseLong(userPk))
+                .orElseThrow(CUserNotFoundException::new);
+    }
+
+}
