+++ src/main/java/com/example/jubging/Service/response/CommonResponse.java
@@ -1,0 +1,14 @@
+package com.example.jubging.Service.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public enum CommonResponse {
+    SUCCESS(0, "성공하였습니다."),
+    FAIL(-1, "실패하였습니다.");
+
+    private int code;
+    private String msg;
+}
