--- src/main/java/com/example/jubging/advice/exception/CEmailLoginFailedException.java
@@ -1,16 +1,0 @@
-package com.example.jubging.advice.exception;
-
-// 로그인 예외처리
-public class CEmailLoginFailedException extends RuntimeException {
-    public CEmailLoginFailedException() {
-        super();
-    }
-
-    public CEmailLoginFailedException(String message) {
-        super(message);
-    }
-
-    public CEmailLoginFailedException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
