--- src/main/java/com/example/jubging/advice/exception/CUserNotFoundException.java
@@ -1,16 +1,0 @@
-package com.example.jubging.advice.exception;
-
-public class CUserNotFoundException extends RuntimeException{
-
-    public CUserNotFoundException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public CUserNotFoundException(String message) {
-        super(message);
-    }
-
-    public CUserNotFoundException() {
-        super();
-    }
-}
