--- src/main/java/com/example/jubging/model/dto/TokenDTO.java
@@ -1,16 +1,0 @@
-package com.example.jubging.model.dto;
-
-import lombok.*;
-import org.springframework.stereotype.Service;
-
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class TokenDTO {
-    private String grantType;
-    private String accessToken;
-    private String refreshToken;
-    private Long accessTokenExpireDate;
-}
