--- src/main/java/com/example/jubging/model/entity/Gender.java
@@ -1,8 +1,0 @@
-package com.example.jubging.model.entity;
-
-public enum Gender {
-// Male - 0
-// Female - 1
-// None - 2
-    MALE, FEMALE, NONE
-}
