--- src/main/java/com/example/jubging/model/entity/RefreshToken.java
@@ -1,36 +1,0 @@
-package com.example.jubging.model.entity;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-@Entity
-@Table(name = "refresh_token")
-@Getter
-@NoArgsConstructor
-public class RefreshToken {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private Long tokenKey;
-
-    @Column(nullable = false)
-    private String token;
-
-    public RefreshToken updateToken(String token) {
-        this.token = token;
-        return this;
-    }
-
-    @Builder
-    public RefreshToken(Long tokenKey, String token) {
-        this.tokenKey = tokenKey;
-        this.token = token;
-    }
-
-}
