--- src/main/java/com/example/jubging/model/response/SingleResult.java
@@ -1,10 +1,0 @@
-package com.example.jubging.model.response;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class SingleResult<T> extends CommonResult{
-    private T data;
-}
