--- src/main/java/com/example/jubging/service/CustomUserDetailService.java
@@ -1,26 +1,0 @@
-package com.example.jubging.service;
-
-import com.example.jubging.advice.exception.CUserNotFoundException;
-import com.example.jubging.model.response.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-
-@RequiredArgsConstructor
-@Service
-public class CustomUserDetailService implements UserDetailsService {
-
-    private final UserRepository userRepository;
-
-    @Override
-    @Transactional(readOnly = true)
-    public UserDetails loadUserByUsername(String userPk) throws UsernameNotFoundException {
-        return userRepository.findById(Long.parseLong(userPk))
-                .orElseThrow(CUserNotFoundException::new);
-    }
-
-}
