--- src/main/java/com/example/jubging/service/response/CommonResponse.java
@@ -1,14 +1,0 @@
-package com.example.jubging.service.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public enum CommonResponse {
-    SUCCESS(0, "성공하였습니다."),
-    FAIL(-1, "실패하였습니다.");
-
-    private int code;
-    private String msg;
-}
