+++ src/main/resources/application.yml
@@ -1,46 +1,7 @@
-## local DB ??
-## mysql ??
-#spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-#spring.datasource.url=jdbc:mysql://localhost:3306/jubging?serverTimezone=UTC&characterEncoding=UTF-8
-#spring.datasource.username=root
-#spring.datasource.password=100412
-#
-## ddl ??? ??????? ?? ?? ??
-#spring.jpa.generate-ddl=true
-#spring.jpa.hibernate.ddl-auto=update
-#
-
-#logging:
-#  level:
-#    root: warn
-#    com.restApi.restApiSpringBootApp: info
-#  file:
-#    path: /home/ec2-user/api/log
-
-#  logback:
-#    rollingpolicy:
-#      max-history: 7
-
 server:
   port: 8080
 
 spring:
-  datasource:
-    url: jdbc:mysql://localhost:3306/jubging?serverTimezone=UTC&characterEncoding=UTF-8
-    driver-class-name: com.mysql.cj.jdbc.Driver
-    username: root
-    password: 108256
-  jpa:
-    database-platform: org.hibernate.dialect.MySQL5InnoDBDialect
-    properties.hibernate:
-#      hbm2ddl.auto: none
-      format_sql: true
-    generate-ddl: true
-    show-sql: false
-    open-in-view: false
-    hibernate:
-      ddl-auto: create
-  url:
-    base: "deployServerBaseUrl"
   profiles:
+    active: dev
     include: jwt
